refactor(review): remove redundant casts, qualifiers, and string prefixes (InspectCode)
Second InspectCode pass (clean isolated worktree). Span-precise deletions, build-validated
(full solution: 0 warnings / 0 errors):
- RedundantCast x10 (e.g. (IFactoryCacheStore), (byte[]), (IHeadlessDbContext))
- RedundantNameQualifier x3 (System.Collections.Concurrent., Amazon.SQS.Model.)
- RedundantStringInterpolation x6 ($ on holeless strings)
- RedundantSuppressNullableWarningExpression x5 (remaining redundant !)
- RedundantVerbatimStringPrefix x4 (remaining redundant @)

+++ src/Headless.Checks/Argument.IsInEnum.cs
@@ -49,9 +49,7 @@ public static partial class Argument
     )
         where T : struct, Enum
     {
-        if (
-            Enum.IsDefined(typeof(T), argument) || _IsValidFlagsCombination(typeof(T), unchecked((ulong)(long)argument))
-        )
+        if (Enum.IsDefined(typeof(T), argument) || _IsValidFlagsCombination(typeof(T), unchecked((ulong)argument)))
         {
             return argument;
         }
